nix: allow the use of local sources
This change allows to run "nix-shell --arg nospdk true" such that
you can have a local copy within spdk-sys/spdk and make changes
there without the need for updating the remote. (see README)

Also, mayastor/default.nix now makes use of sources.json, as well
as the rust overlay.

As a bonus, some small changes from usize to u64 are part of this change
as bindgen erroneously converted them, by default, to u64. This
will reduce the diff for the upgrade that is coming.

+++ mayastor/src/subsys/mod.rs
@@ -70,7 +70,7 @@ impl MayastorSubsystem {
             spdk_json_write_val_raw(
                 w,
                 data.as_ptr() as *const _,
-                data.as_bytes().len(),
+                data.as_bytes().len() as u64,
             );
         }
     }
